# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bütçe API",
        "version": "1.0"
    },
    "paths": {
...
        "/employeeTypes/findAll": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Find all employee types",
                "operationId": "employeeTypes_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeTypesFindAllResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
...
        "/files/findOne/{fileId}": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Find ",
                "operationId": "files_findOne",
                "parameters": [
                    {
                        "name": "fileId",
                        "in": "path",
                        "description": "File id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tableName",
                        "in": "query",
                        "description": "File tableName",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation (File content returns to Base64 string)"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmployeeTypesFindAllResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}